+++ tasks/integration/parse-number.js
@@ -3,6 +3,6 @@ const test = require('./test');
 
 
 const st = Streams.ofString('12');
-const parsing = N.numberLiteral.thenLeft(F.eos).parse(st);
+const parsing = N.numberLiteral().thenLeft(F.eos()).parse(st);
 
 test("Simple number parsing", 12, parsing.value);